Address review: uint mode flag, atol examples, UMA offload and cache notes
- adam_w_mode reaches the Metal kernel as constant uint& instead of a
  float flag; the shader binding marshals Python ints natively.
- Spell out what the ulp-scaled test tolerance evaluates to against the
  measured implementation agreement.
- Docs: on unified memory offload is not about capacity; it moves the
  optimizer state out of Metal's GPU working-set budget and the step
  onto the CPU cores.
- MetalOpBuilder: note why there is no torch-extensions disk cache
  (millisecond runtime compiles; Metal keeps its own pipeline cache).

+++ csrc/mps/fused_adam.metal
@@ -20,7 +20,7 @@ kernel void fused_adam(device T* param [[buffer(0)]],
                        constant float& weight_decay [[buffer(8)]],
                        constant float& bias_correction1 [[buffer(9)]],
                        constant float& bias_correction2 [[buffer(10)]],
-                       constant float& adam_w_mode [[buffer(11)]],
+                       constant uint& adam_w_mode [[buffer(11)]],
                        uint i [[thread_position_in_grid]])
 {
     float g = float(grad[i]);
@@ -29,12 +29,12 @@ kernel void fused_adam(device T* param [[buffer(0)]],
     float v = float(exp_avg_sq[i]);
 
     // L2 mode folds weight decay into the gradient; AdamW mode applies it to the parameter.
-    if (adam_w_mode == 0.0f) { g += weight_decay * p; }
+    if (adam_w_mode == 0) { g += weight_decay * p; }
     m = beta1 * m + (1.0f - beta1) * g;
     v = beta2 * v + (1.0f - beta2) * g * g;
     float denom = sqrt(v / bias_correction2) + epsilon;
     float update = (m / bias_correction1) / denom;
-    if (adam_w_mode != 0.0f) { update += weight_decay * p; }
+    if (adam_w_mode != 0) { update += weight_decay * p; }
 
     param[i] = T(p - lr * update);
     exp_avg[i] = T(m);
@@ -44,15 +44,15 @@ kernel void fused_adam(device T* param [[buffer(0)]],
 template [[host_name("fused_adam_float")]] kernel void fused_adam<float>(
     device float*, device const float*, device float*, device float*,
     constant float&, constant float&, constant float&, constant float&, constant float&,
-    constant float&, constant float&, constant float&, uint);
+    constant float&, constant float&, constant uint&, uint);
 template [[host_name("fused_adam_half")]] kernel void fused_adam<half>(
     device half*, device const half*, device half*, device half*,
     constant float&, constant float&, constant float&, constant float&, constant float&,
-    constant float&, constant float&, constant float&, uint);
+    constant float&, constant float&, constant uint&, uint);
 // bfloat is a Metal 3.1 type (macOS 14); older toolchains still get the float and half kernels.
 #if __METAL_VERSION__ >= 310
 template [[host_name("fused_adam_bfloat")]] kernel void fused_adam<bfloat>(
     device bfloat*, device const bfloat*, device bfloat*, device bfloat*,
     constant float&, constant float&, constant float&, constant float&, constant float&,
-    constant float&, constant float&, constant float&, uint);
+    constant float&, constant float&, constant uint&, uint);
 #endif
